LayoutTests/platform/wpe/imported/w3c/web-platform-tests/html/browsers/windows/auxiliary-browsing-contexts/opener-noopener-expected.txt

CONSOLE MESSAGE: QuotaExceededError: The quota has been exceeded.

Harness Error (TIMEOUT), message = null

TIMEOUT Auxiliary browsing context created via `window.open` setting `noopener` should report `window.opener` `null` Test timed out
